+++ kakaobase/src/features/feeds/posts/hooks/usePostEditorForm.tsx
@@ -1,11 +1,8 @@
 import postToS3 from '@/entities/images/api/imageS3';
 import { postPost } from '../../api/post';
 import { queryClient } from '@/shared/api/queryClient';
-import { postSchema } from '../schemas/postSchema';
-import { zodResolver } from '@hookform/resolvers/zod';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
-import { z } from 'zod';
 import { useToast } from '@/shared/hooks/ToastContext';
 import { useUserStore } from '@/entities/users/stores/userStore';
 import { feedQueries } from '../../api/feedQueries';
@@ -13,7 +10,11 @@ import useRoutings from '@/shared/hooks/useRoutings';
 import { accountListQueries } from '@/features/account/api/accountListQueries';
 import { accountInfoQueries } from '@/features/account/api/accountInfoQueries';
 
-export type NewPostData = z.infer<typeof postSchema>;
+export type NewPostData = {
+  content?: string;
+  youtubeUrl?: string;
+  imageFile?: File;
+};
 
 export const usePostEditorForm = () => {
   const { showToast } = useToast();
@@ -22,7 +23,15 @@ export const usePostEditorForm = () => {
   const { goMain } = useRoutings();
 
   const methods = useForm<NewPostData>({
-    resolver: zodResolver(postSchema),
+    resolver: async (...args) => {
+      const [{ z }, { zodResolver }, { postSchema }] = await Promise.all([
+        import('zod'),
+        import('@hookform/resolvers/zod'),
+        import('../schemas/postSchema'),
+      ]);
+      const schema = z.object({ postSchema });
+      return zodResolver(schema)(...args);
+    },
     mode: 'all',
     reValidateMode: 'onChange',
     defaultValues: {
