+++ kakaobase/src/features/feeds/posts/ui/HelperText.tsx
@@ -1,0 +1,3 @@
+export default function HelperText({ errorMessage }: { errorMessage: string }) {
+  return <div className="text-redHeart text-xs h-4">{errorMessage}</div>;
+}
